+++ src/components/domain/editprofile/company/index.ts
@@ -1,0 +1,1 @@
+export { default as Company } from './Company'
