+++ src/components/domain/editprofile/experience/index.ts
@@ -1,0 +1,1 @@
+export { default as Experience } from './Experience'
