+++ src/components/domain/editprofile/jobgroup/index.ts
@@ -1,0 +1,1 @@
+export { default as JobGroup } from './JobGroup'
