+++ src/components/domain/editprofile/nickname/index.ts
@@ -1,0 +1,1 @@
+export { default as NickName } from './NickName'
