+++ src/components/domain/editprofile/profileimg/index.ts
@@ -1,0 +1,1 @@
+export { default as ProfileImg } from './ProfileImg'
